requirements.txt
fastapi
uvicorn
sqlmodel[all]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pymysql